@@ -4,8 +4,8 @@
   "description": "UI for rendering Node Graphs.",
   "main": "lib/index.js",
   "peerDependencies": {
-    "react": ">=0.13.0 <0.15.0 || ^0.14.0-beta3",
-    "react-dom": ">=0.14.0 <0.15.0 || ^0.14.0-beta3"
+    "react": "^16.9.0",
+    "react-dom": "^16.9.0"
   },
   "devDependencies": {
     "babel": "^5.6.14",
@@ -13,15 +13,15 @@
     "babel-core": "^5.6.18",
     "babel-eslint": "^4.1.3",
     "babel-loader": "^5.3.1",
-    "react": "^0.14.0",
+    "react": "^16.9.0",
     "react-hot-loader": "^1.2.8",
     "webpack": "^1.10.1",
     "webpack-dev-server": "^1.10.1"
   },
   "dependencies": {
-    "react-dom": "^0.14.0",
+    "react-dom": "^16.9.0",
     "react-draggable": "^3.3.2",
-    "react-onclickoutside": "^5.11.1"
+    "react-onclickoutside": "^6.9.0"
   },
   "scripts": {
     "build": "babel src --out-dir lib && webpack --config webpack.prod.config.js",
