@@ -22,6 +22,10 @@ var _reactOnclickoutside = require('react-onclickoutside');
 
 var _reactOnclickoutside2 = _interopRequireDefault(_reactOnclickoutside);
 
+var _reactDraggable = require('react-draggable');
+
+var _reactDraggable2 = _interopRequireDefault(_reactDraggable);
+
 var _NodeInputList = require('./NodeInputList');
 
 var _NodeInputList2 = _interopRequireDefault(_NodeInputList);
@@ -29,8 +33,6 @@ var _NodeInputList2 = _interopRequireDefault(_NodeInputList);
 var _NodeOutputList = require('./NodeOutputList');
 
 var _NodeOutputList2 = _interopRequireDefault(_NodeOutputList);
-
-var Draggable = require('react-draggable');
 
 var Node = (function (_React$Component) {
   _inherits(Node, _React$Component);
@@ -76,7 +78,7 @@ var Node = (function (_React$Component) {
     }
   }, {
     key: 'handleClick',
-    value: function handleClick(e) {
+    value: function handleClick() {
       this.setState({ selected: true });
       if (this.props.onNodeSelect) {
         this.props.onNodeSelect(this.props.nid);
@@ -107,7 +109,7 @@ var Node = (function (_React$Component) {
             _this.handleClick(e);
           } },
         _react2['default'].createElement(
-          Draggable,
+          _reactDraggable2['default'],
           {
             start: { x: this.props.pos.x, y: this.props.pos.y },
             handle: '.node-header',
