@@ -1,71 +1,43 @@
 "use strict";
 
 Object.defineProperty(exports, "__esModule", {
-	value: true
+    value: true
 });
 
-var _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();
-
-var _get = function get(_x, _x2, _x3) { var _again = true; _function: while (_again) { var object = _x, property = _x2, receiver = _x3; _again = false; if (object === null) object = Function.prototype; var desc = Object.getOwnPropertyDescriptor(object, property); if (desc === undefined) { var parent = Object.getPrototypeOf(object); if (parent === null) { return undefined; } else { _x = parent; _x2 = property; _x3 = receiver; _again = true; desc = parent = undefined; continue _function; } } else if ("value" in desc) { return desc.value; } else { var getter = desc.get; if (getter === undefined) { return undefined; } return getter.call(receiver); } } };
-
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { "default": obj }; }
-
-function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
-
-function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
 
 var _react = require('react');
 
 var _react2 = _interopRequireDefault(_react);
 
-var NodeOutputListItem = (function (_React$Component) {
-	_inherits(NodeOutputListItem, _React$Component);
+var NodeOutputListItem = function NodeOutputListItem(_ref) {
+    var onMouseDown = _ref.onMouseDown;
+    var index = _ref.index;
+    var item = _ref.item;
 
-	function NodeOutputListItem() {
-		_classCallCheck(this, NodeOutputListItem);
+    var handleMouseDown = (0, _react.useCallback)(function (e) {
+        e.stopPropagation();
+        e.preventDefault();
 
-		_get(Object.getPrototypeOf(NodeOutputListItem.prototype), "constructor", this).apply(this, arguments);
-	}
+        onMouseDown(index);
+    }, []);
 
-	_createClass(NodeOutputListItem, [{
-		key: "onMouseDown",
-		value: function onMouseDown(e) {
-			e.stopPropagation();
-			e.preventDefault();
+    var noop = (0, _react.useCallback)(function (e) {
+        e.stopPropagation();
+        e.preventDefault();
+    }, []);
 
-			this.props.onMouseDown(this.props.index);
-		}
-	}, {
-		key: "noop",
-		value: function noop(e) {
-			e.stopPropagation();
-			e.preventDefault();
-		}
-	}, {
-		key: "render",
-		value: function render() {
-			var _this = this;
+    return _react2["default"].createElement(
+        "li",
+        { onMouseDown: handleMouseDown },
+        _react2["default"].createElement(
+            "a",
+            { href: "#", onClick: noop },
+            item.name,
+            _react2["default"].createElement("i", { className: "fa fa-circle-o" })
+        )
+    );
+};
 
-			return _react2["default"].createElement(
-				"li",
-				{ onMouseDown: function (e) {
-						return _this.onMouseDown(e);
-					} },
-				_react2["default"].createElement(
-					"a",
-					{ href: "#", onClick: function (e) {
-							return _this.noop(e);
-						} },
-					this.props.item.name,
-					" ",
-					_react2["default"].createElement("i", { className: "fa fa-circle-o" })
-				)
-			);
-		}
-	}]);
-
-	return NodeOutputListItem;
-})(_react2["default"].Component);
-
-exports["default"] = NodeOutputListItem;
+exports["default"] = _react2["default"].memo(NodeOutputListItem);
 module.exports = exports["default"];
